taito/lkage.cpp: consolidated driver in single file
